move cli.go into commands package
This lets us import the cli and use it to run cts similar to the e2e
tests, runs the full app, but keeping it in process. This is
particularly useful for benchmarking, so we can profile the benchmarks.
